+++ converter/smr2rdox/CMakeLists.txt
@@ -74,7 +74,6 @@ SET(CONVERTER_ERROR_FILES
 
 SET(CONVERTER_LEXER_FILES
 	rdoparser_lexer.cpp
-	rdoparser_lexer-inl.h
 	rdoparser_lexer.h
 )
 
@@ -177,7 +176,6 @@ SET(RUNTIME_FILES
 	runtime/rdo_ie.h
 	runtime/rdo_logic_dptfree.cpp
 	runtime/rdo_logic_dptfree.h
-	runtime/rdo_logic_dptfree-inl.h
 )
 
 SET(TYPE_FILES
@@ -209,7 +207,6 @@ SET(UPDATE_FILES
 SET(VALUE_FILES
 	rdo_value.cpp
 	rdo_value.h
-	rdo_value-inl.h
 )
 
 SET(VALUE_ARRAY_FILES
