+++ converter/smr2rdox/runtime/rdo_logic_dptfree.cpp
@@ -13,3 +13,17 @@
 // ----------------------------------------------------------------------- SYNOPSIS
 #include "converter/smr2rdox/runtime/rdo_logic_dptfree.h"
 // --------------------------------------------------------------------------------
+
+OPEN_RDO_RUNTIME_NAMESPACE
+
+// --------------------------------------------------------------------------------
+// -------------------- RDODPTFree
+// --------------------------------------------------------------------------------
+RDODPTFree::RDODPTFree(const LPRDORuntime& pRuntime)
+	: RDOLogicSimple(pRuntime, NULL)
+{}
+
+RDODPTFree::~RDODPTFree()
+{}
+
+CLOSE_RDO_RUNTIME_NAMESPACE
